requirements.txt
cryptography==45.0.7
pyotp==2.9.0
qrcode==8.2
pillow==11.3.0
argon2-cffi==25.1.0
streamlit==1.49.1
streamlit-cookies-manager==0.2.0
